requirements.txt
streamlit
tensorflow==2.10.0
keras==2.10.0
numpy
pillow
gdown
opencv-python-headless